feat!: change how server handles images
Previously, the column in the database was storing a link to a hosted image somewhere. Now, it returns the filename to a local file on the server, located in a folder that the server is statically serving files from. The client can directly grab images from the server now, and uploaded images will go into this folder as well. The reason for this change was to allow the client to actually upload images when  creating a blog post, rather than having to paste a link to an already hosted image.